frontend: helpers: Extract getHeadlampAPIHeaders to own module
Also moved setBackendToken in there.
