+++ frontend/src/components/App/AppContainer.tsx
@@ -2,8 +2,8 @@ import { GlobalStyles } from '@mui/material';
 import { SnackbarProvider } from 'notistack';
 import React from 'react';
 import { BrowserRouter, HashRouter } from 'react-router-dom';
-import { setBackendToken } from '../../helpers';
 import { getBaseUrl } from '../../helpers/getBaseUrl';
+import { setBackendToken } from '../../helpers/getHeadlampAPIHeaders';
 import { isElectron } from '../../helpers/isElectron';
 import Plugins from '../../plugin/Plugins';
 import ReleaseNotes from '../common/ReleaseNotes/ReleaseNotes';
