+++ frontend/src/helpers/getHeadlampAPIHeaders.ts
@@ -1,0 +1,31 @@
+/**
+ * The backend token to use when making API calls from Headlamp when running as an app.
+ * The token is requested from the main process via IPC once the renderer is ready,
+ * and stored for use in the getHeadlampAPIHeaders function below.
+ *
+ * The app also sets HEADLAMP_BACKEND_TOKEN in the headlamp-server environment,
+ * which the server checks to validate requests containing this same token.
+ */
+let backendToken: string | null = null;
+
+/**
+ * Sets the backend token to use when making API calls from Headlamp when running as an app.
+ *
+ * This is not a K8s or OIDC token, but one that protects headlamp-server APIs.
+ */
+export function setBackendToken(token: string | null) {
+  backendToken = import.meta.env.REACT_APP_HEADLAMP_BACKEND_TOKEN || token;
+}
+
+/**
+ * Returns headers for making API calls to the headlamp-server backend.
+ */
+export function getHeadlampAPIHeaders(): { [key: string]: string } {
+  if (backendToken === null) {
+    return {};
+  }
+
+  return {
+    'X-HEADLAMP_BACKEND-TOKEN': backendToken,
+  };
+}
