+++ frontend/src/helpers/index.ts
@@ -2,6 +2,7 @@ import { Cluster } from '../lib/k8s/cluster';
 import { loadClusterSettings, storeClusterSettings } from './clusterSettings';
 import { getAppUrl } from './getAppUrl';
 import { getBaseUrl } from './getBaseUrl';
+import { getHeadlampAPIHeaders } from './getHeadlampAPIHeaders';
 import { isDevMode } from './isDevMode';
 import { isDockerDesktop } from './isDockerDesktop';
 import { isElectron } from './isElectron';
@@ -77,33 +78,6 @@ function loadTableSettings(tableId: string): { id: string; show: boolean }[] {
 
   const settings = JSON.parse(localStorage.getItem(`table_settings.${tableId}`) || '[]');
   return settings;
-}
-
-/**
- * The backend token to use when making API calls from Headlamp when running as an app.
- * The token is requested from the main process via IPC once the renderer is ready,
- * and stored for use in the getHeadlampAPIHeaders function below.
- *
- * The app also sets HEADLAMP_BACKEND_TOKEN in the headlamp-server environment,
- * which the server checks to validate requests containing this same token.
- */
-let backendToken: string | null = null;
-
-export function setBackendToken(token: string | null) {
-  backendToken = import.meta.env.REACT_APP_HEADLAMP_BACKEND_TOKEN || token;
-}
-
-/**
- * Returns headers for making API calls to the headlamp-server backend.
- */
-export function getHeadlampAPIHeaders(): { [key: string]: string } {
-  if (backendToken === null) {
-    return {};
-  }
-
-  return {
-    'X-HEADLAMP_BACKEND-TOKEN': backendToken,
-  };
 }
 
 const exportFunctions = {
